+++ tests/generator/__init__.py
@@ -1,0 +1,1 @@
+"""Code generator unit tests"""